Extracted Search into it's own widget page, extracted repeated code into a widgets file
