+++ lib/views/messages/inbox.dart
@@ -7,6 +7,8 @@ import 'package:amplify_datastore/amplify_datastore.dart';
 // amplify configuration and models
 import '../../models/ModelProvider.dart';
 import '../../models/Messages.dart';
+// Custom Widgets for app
+import './widgets/widgets.dart';
 
 class InboxPage extends StatefulWidget {
   /// Builds the inbox of most recent [messages] per unique conversation.
@@ -24,10 +26,10 @@ class InboxPage extends StatefulWidget {
 
 class _InboxPageState extends State<InboxPage> {
 
+  late StreamSubscription<QuerySnapshot<Messages>> messageStream;
   List<Messages> _messages = [];
   bool _isLoading = true;
-  late StreamSubscription<QuerySnapshot<Messages>> messageStream;
-
+  
   @override
   void initState() {
     super.initState();
@@ -53,40 +55,33 @@ class _InboxPageState extends State<InboxPage> {
   Widget build(BuildContext context) {
     
     List<String?> args = ModalRoute.of(context)!.settings.arguments as List<String?>;
-    final String userName = args[0].toString();
+    final String userName = args[0]!;
 
     if(_isLoading) getMessageStream(userName);
     
-    return (
-      Scaffold(
-        appBar: AppBar(
-          leading: const BackButton(),
-          title: const Text('Inbox'),
-          backgroundColor: Theme.of(context).primaryColor,
-          centerTitle: true,
-        ),
+    return Scaffold(
+        appBar: appBar('Inbox', context),
         body: _isLoading
-            ? Center(child: CircularProgressIndicator(
-                color: Theme.of(context).primaryColor))
+            ? progressIndicator(context)
             : _messages.isNotEmpty
-                ? Column(
-                    children: [
-                      Expanded(
-                        flex: 8,
-                        child: InboxList(
-                            messages: filterRecentMessagesByGroup(_messages),
-                            dataStore: widget.dataStore,
-                            userName: userName)),
-                      Expanded(
-                          flex: 2,
-                          child: Container(
-                            color: Theme.of(context).primaryColor,
-                          ))
-                    ],
-                  )
-                : const Center(child: Text('No messages'))));
+              ? Column(
+                children: [
+                  Expanded(
+                    flex: 8,
+                    child: InboxList(
+                        messages: filterRecentMessagesByGroup(_messages),
+                        dataStore: widget.dataStore,
+                        userName: userName)),
+                  Expanded(
+                      flex: 2,
+                      child: Container(
+                        color: Theme.of(context).primaryColor,
+                      ))
+                  ]
+                )
+              : const Center(child: Text('No messages')));
+    }
   }
-}
 
 class InboxList extends StatelessWidget {
   /// Builds the inbox items into a [ListView] for browsing by the user.
@@ -152,14 +147,13 @@ class InboxItem extends StatelessWidget {
         child: (
           ListTile(
             title: message.customer!= userName 
-                  ? userNameAndMessageText(message.customer, message.text, context)
-                  : userNameAndMessageText(message.host, message.text, context),
+              ? userNameAndMessageText(message.customer, message.text, context)
+              : userNameAndMessageText(message.host, message.text, context),
             trailing: formattedDate(message, userName, context),
             onTap: () => {
-                  Navigator.pushNamed(context, '/msgDetail',
-                      arguments: [userName, message.sale, message.customer])
-            },
-          ))
+              Navigator.pushNamed(context, '/msgDetail',
+                  arguments: [userName, message.sale, message.customer])
+            }))
       )
     );
   }
@@ -234,6 +228,7 @@ List<Messages> filterRecentMessagesByGroup(List<Messages> messages) {
   bool flag = true;
 
   for (int i = 0; i < messages.length; i++) {
+
     if(formattedMessages.isEmpty) formattedMessages.add(messages[i]);
 
     for (int j = 0; j < formattedMessages.length; j++) {
@@ -249,6 +244,7 @@ List<Messages> filterRecentMessagesByGroup(List<Messages> messages) {
             break;
         }
     }
+    
     if(flag) formattedMessages.add(messages[i]);
     flag = true;
   }
